refactor: consolidate ActivationStore into FileActivationStateStore
Pre-Phase 4 store consolidation — eliminates the pass-through wrapper
class and moves security event logging into the file-backed store.
All consumers now reference IActivationStateStore interface, enabling
clean database backend swap in Phase 4.

Session-scoped log filtering hardened: query_logs rejects cross-session
queries when a session context is active. Strengthened Phase 3.4
isolation test assertions for Gatekeeper confirmation and cleanup
verification.

# Conflicts:
#	src/handlers/ElementCRUDHandler.ts
#	src/services/ActivationStore.ts
#	src/state/IActivationStateStore.ts
