+++ tests/integration/mcp-aql/activation-persistence.test.ts
@@ -21,17 +21,8 @@ import os from 'os';
 import { DollhouseMCPServer } from '../../../src/index.js';
 import { DollhouseContainer } from '../../../src/di/Container.js';
 import { MCPAQLHandler } from '../../../src/handlers/mcp-aql/MCPAQLHandler.js';
-import { ActivationStore } from '../../../src/services/ActivationStore.js';
+import type { IActivationStateStore } from '../../../src/state/IActivationStateStore.js';
 import { FileActivationStateStore } from '../../../src/state/FileActivationStateStore.js';
-
-/**
- * Helper: create an ActivationStore with the old (fileOps, stateDir, sessionId) pattern.
- * Issue #1945: ActivationStore now takes IActivationStateStore directly.
- */
-function createActivationStore(fileOps: any, stateDir?: string, sessionId?: string) {
-  const backingStore = new FileActivationStateStore(fileOps, stateDir, sessionId);
-  return new ActivationStore(backingStore);
-}
 import { createPortfolioTestEnvironment, preConfirmAllOperations, type PortfolioTestEnvironment } from '../../helpers/portfolioTestHelper.js';
 import type { OperationResult } from '../../../src/handlers/mcp-aql/types.js';
 
@@ -40,7 +31,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
   let container: DollhouseContainer;
   let server: DollhouseMCPServer;
   let mcpAqlHandler: MCPAQLHandler;
-  let activationStore: ActivationStore;
+  let activationStore: IActivationStateStore;
 
   // Save env vars we modify
   let originalSessionId: string | undefined;
@@ -61,7 +52,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
     preConfirmAllOperations(container);
 
     mcpAqlHandler = container.resolve<MCPAQLHandler>('mcpAqlHandler');
-    activationStore = container.resolve<ActivationStore>('ActivationStore');
+    activationStore = container.resolve<IActivationStateStore>('ActivationStore');
   });
 
   afterEach(async () => {
@@ -83,8 +74,8 @@ describe('Activation Persistence Integration (Issue #598)', () => {
 
   describe('DI wiring verification', () => {
     it('should have ActivationStore registered as singleton', () => {
-      const store1 = container.resolve<ActivationStore>('ActivationStore');
-      const store2 = container.resolve<ActivationStore>('ActivationStore');
+      const store1 = container.resolve<IActivationStateStore>('ActivationStore');
+      const store2 = container.resolve<IActivationStateStore>('ActivationStore');
       expect(store1).toBe(store2);
     });
 
@@ -255,7 +246,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
       try {
         // Create a store with known state dir
         const fileOps = container.resolve<any>('FileOperationsService');
-        const store1 = createActivationStore(fileOps, stateDir);
+        const store1 = new FileActivationStateStore(fileOps, stateDir);
         await store1.initialize();
 
         // Record some activations
@@ -271,7 +262,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
         expect(files.some(f => f.startsWith('activations-'))).toBe(true);
 
         // Create a NEW store instance and load from the same directory
-        const store2 = createActivationStore(fileOps, stateDir);
+        const store2 = new FileActivationStateStore(fileOps, stateDir);
         await store2.initialize();
 
         // Verify round-trip
@@ -303,7 +294,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
         const fileOps = container.resolve<any>('FileOperationsService');
 
         // Store 1: activate two skills, deactivate one
-        const store1 = createActivationStore(fileOps, stateDir);
+        const store1 = new FileActivationStateStore(fileOps, stateDir);
         await store1.initialize();
         store1.recordActivation('skill', 'keep-me');
         store1.recordActivation('skill', 'remove-me');
@@ -313,7 +304,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
         await new Promise(resolve => setTimeout(resolve, 200));
 
         // Store 2: load and verify
-        const store2 = createActivationStore(fileOps, stateDir);
+        const store2 = new FileActivationStateStore(fileOps, stateDir);
         await store2.initialize();
 
         const skills = store2.getActivations('skill');
@@ -334,14 +325,14 @@ describe('Activation Persistence Integration (Issue #598)', () => {
 
         // Session A: activate skill-a
         process.env.DOLLHOUSE_SESSION_ID = 'session-alpha';
-        const storeA = createActivationStore(fileOps, stateDir);
+        const storeA = new FileActivationStateStore(fileOps, stateDir);
         await storeA.initialize();
         storeA.recordActivation('skill', 'skill-a');
         await new Promise(resolve => setTimeout(resolve, 200));
 
         // Session B: activate skill-b (different skill)
         process.env.DOLLHOUSE_SESSION_ID = 'session-beta';
-        const storeB = createActivationStore(fileOps, stateDir);
+        const storeB = new FileActivationStateStore(fileOps, stateDir);
         await storeB.initialize();
         storeB.recordActivation('skill', 'skill-b');
         await new Promise(resolve => setTimeout(resolve, 200));
@@ -353,13 +344,13 @@ describe('Activation Persistence Integration (Issue #598)', () => {
 
         // Verify data isolation — reload each
         process.env.DOLLHOUSE_SESSION_ID = 'session-alpha';
-        const reloadA = createActivationStore(fileOps, stateDir);
+        const reloadA = new FileActivationStateStore(fileOps, stateDir);
         await reloadA.initialize();
         expect(reloadA.getActivations('skill').some(a => a.name === 'skill-a')).toBe(true);
         expect(reloadA.getActivations('skill').some(a => a.name === 'skill-b')).toBe(false);
 
         process.env.DOLLHOUSE_SESSION_ID = 'session-beta';
-        const reloadB = createActivationStore(fileOps, stateDir);
+        const reloadB = new FileActivationStateStore(fileOps, stateDir);
         await reloadB.initialize();
         expect(reloadB.getActivations('skill').some(a => a.name === 'skill-b')).toBe(true);
         expect(reloadB.getActivations('skill').some(a => a.name === 'skill-a')).toBe(false);
@@ -372,7 +363,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
   });
 
   describe('Stale element pruning', () => {
-    it('should skip stale activations during restore when element no longer exists', async () => {
+    it('should load stale activations into memory (pruning happens in Container.restoreActivations)', async () => {
       const stateDir = await fs.mkdtemp(path.join(os.tmpdir(), 'activation-store-stale-'));
 
       try {
@@ -397,7 +388,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
         );
 
         // Load the store — should succeed (loads in-memory)
-        const store = createActivationStore(fileOps, stateDir);
+        const store = new FileActivationStateStore(fileOps, stateDir);
         await store.initialize();
 
         // The store should have the activation loaded
@@ -423,7 +414,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
           'NOT_VALID_JSON{{{corrupted'
         );
 
-        const store = createActivationStore(fileOps, stateDir);
+        const store = new FileActivationStateStore(fileOps, stateDir);
         await store.initialize(); // Should not throw
 
         // Should start with empty activations
@@ -455,7 +446,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
           JSON.stringify(futureState, null, 2)
         );
 
-        const store = createActivationStore(fileOps, stateDir);
+        const store = new FileActivationStateStore(fileOps, stateDir);
         await store.initialize();
 
         // Version mismatch should cause it to start fresh
@@ -474,7 +465,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
 
       try {
         const fileOps = container.resolve<any>('FileOperationsService');
-        const store = createActivationStore(fileOps, stateDir);
+        const store = new FileActivationStateStore(fileOps, stateDir);
         await store.initialize();
 
         expect(store.isEnabled()).toBe(false);
@@ -501,7 +492,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
 
       try {
         const fileOps = container.resolve<any>('FileOperationsService');
-        const store = createActivationStore(fileOps, stateDir);
+        const store = new FileActivationStateStore(fileOps, stateDir);
         await store.initialize();
 
         store.recordActivation('skill', 'clear-test-skill');
@@ -520,7 +511,7 @@ describe('Activation Persistence Integration (Issue #598)', () => {
         expect(store.getActivations('memory')).toHaveLength(0);
 
         // Reload from disk — should be empty
-        const store2 = createActivationStore(fileOps, stateDir);
+        const store2 = new FileActivationStateStore(fileOps, stateDir);
         await store2.initialize();
         expect(store2.getActivations('skill')).toHaveLength(0);
         expect(store2.getActivations('memory')).toHaveLength(0);
